Some refinements in the Sink and Buffe classes.

+++ tests/test_buf.py
@@ -18,10 +18,11 @@
 
 import io
 import os
+from pathlib import Path
 import pytest
 
 from baldaquin import logger, BALDAQUIN_DATA
-from baldaquin.buf import FIFO, CircularBuffer, Sink
+from baldaquin.buf import FIFO, CircularBuffer, WriteMode, Sink
 from baldaquin.pkt import Layout, Format, packetclass, FixedSizePacketBase
 
 
@@ -36,9 +37,9 @@ class Packet(FixedSizePacketBase):
     packet_id: Format.UNSIGNED_SHORT
 
     def as_text(self):
+        """Simple text formatting function.
         """
-        """
-        return 'Packet\n'
+        return f'{self.header}, {self.packet_id}\n'
 
 
 def _test_buffer_base(buffer_class, num_packets: int = 10, **kwargs):
@@ -66,42 +67,65 @@ def test_circular_buffer():
     _test_buffer_base(CircularBuffer)
 
 
+def _scrap_file(file_path: Path) -> Path:
+    """Make sure that a given file does not exist, and remove if it does.
+    """
+    if file_path.exists():
+        os.remove(file_path)
+    return file_path
+
+
 def test_sink_contextmanager():
     """Test the context manager sink protocol.
     """
-    file_path = BALDAQUIN_DATA / 'sink.dat'
-    try:
-        os.remove(file_path)
-    except FileNotFoundError:
-        pass
-    sink = Sink(file_path)
+    file_path = _scrap_file(BALDAQUIN_DATA / 'sink.dat')
+    # Create the sink.
+    sink = Sink(file_path, WriteMode.BINARY)
+    # Open the sink.
     with sink.open() as output_file:
         assert isinstance(output_file, io.IOBase)
+    # Check that creating another sink with the same file path raises an exception.
+    with pytest.raises(FileExistsError) as info:
+        sink = Sink(file_path, WriteMode.BINARY)
+    logger.info(info.value)
+    # Cleanup.
+    os.remove(file_path)
+
+
+def test_sink_header():
+    """Test the header for a text sink.
+    """
+    file_path = _scrap_file(BALDAQUIN_DATA / 'sink.txt')
+    header = '#This is a file header.\n'
+    # Create the sink---this should write the header into the output file...
+    _ = Sink(file_path, WriteMode.TEXT, header=header)
+    # ... open the output file and make sure the content is correct.
+    with open(file_path, 'r') as input_file:
+        assert input_file.read() == header
+    # Cleanup.
     os.remove(file_path)
 
 
 def test_buffer_flush(num_packets=10):
+    """Test the full flush mechanism with multiple sinks.
     """
-    """
+    binary_file_path = _scrap_file(BALDAQUIN_DATA / 'sink.dat')
+    text_file_path = _scrap_file(BALDAQUIN_DATA / 'sink.txt')
+    # Create a buffer and assign two sinks.
     buffer = CircularBuffer()
-    file_path_b = BALDAQUIN_DATA / 'sink.dat'
-    file_path_t = BALDAQUIN_DATA / 'sink.txt'
-    try:
-        os.remove(file_path_b)
-    except FileNotFoundError:
-        pass
-    try:
-        os.remove(file_path_t)
-    except FileNotFoundError:
-        pass
-
-    buffer.add_sink(file_path_b)
-    buffer.add_sink(file_path_t, 't', Packet.as_text)
-
+    buffer.add_sink(binary_file_path, WriteMode.BINARY)
+    buffer.add_sink(text_file_path, WriteMode.TEXT, Packet.as_text, '#File header\n')
+    # Fill the buffer with some packets.
     for i in range(num_packets):
         packet = Packet(Packet.header, i)
         buffer.put(packet)
-
+    # At this point the buffer should have the target number of packets.
+    assert buffer.size() == num_packets
     buffer.flush()
-    #os.remove(file_path_b)
-    #os.remove(file_path_t)
+    # And now it should be empty.
+    assert buffer.size() == 0
+    with open(text_file_path, 'r') as input_file:
+        logger.info(input_file.read())
+    # Cleanup
+    os.remove(binary_file_path)
+    os.remove(text_file_path)
